requirements.txt
# ÖNIKA LI 依赖配置
python-telegram-bot==20.7
requests==2.31.0
urllib3==2.1.0
beautifulsoup4==4.12.2
PyYAML==6.0.1
aiohttp==3.9.1
aiohttp-retry==2.8.3
python-dotenv==1.0.0
schedule==1.2.0
pillow==10.2.0
openai>=1.0.0
anthropic>=0.18.0
